requirements.txt
appdirs==1.4.4
attrs==20.2.0
cached-property==1.5.2
certifi==2020.6.20
chardet==3.0.4
cx-Oracle==8.0.1
defusedxml==0.6.0
Django==2.1.15
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
django-pyodbc-azure-2019==2.1.0.0
djangorestframework==3.9.4
et-xmlfile==1.0.1
flake8==3.6.0
idna==2.10
isodate==0.6.0
jdcal==1.4.1
lxml==4.5.2
mccabe==0.6.1
numpy==1.19.2
openpyxl==3.0.5
pandas==1.1.2
pycodestyle==2.4.0
pyflakes==2.0.0
pyodbc==4.0.30
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
xlrd==1.2.0
zeep==3.4.0